LayoutTests/crypto/webkitSubtle/rsa-oaep-generate-non-extractable-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test that even when non-extractability is requested, the public key is still extractable.");

jsTestIsAsync = true;

var algorithmKeyGen = {
    name: "RSA-OAEP",
    // RsaKeyGenParams
    modulusLength: 2048,
    publicExponent: new Uint8Array([0x01, 0x00, 0x01]),  // Equivalent to 65537
};
var nonExtractable = false;

debug("Generating a key pair...");
crypto.webkitSubtle.generateKey(algorithmKeyGen, nonExtractable, []).then(function(result) {
    keyPair = result;
    shouldBe("keyPair.toString()", "'[object Object]'");
    shouldBe("keyPair.publicKey.type", "'public'");
    shouldBe("keyPair.publicKey.extractable", "true");
    shouldBe("keyPair.publicKey.algorithm.name", "'RSA-OAEP'");
    shouldBe("keyPair.publicKey.algorithm.modulusLength", "2048");
    shouldBe("bytesToHexString(keyPair.publicKey.algorithm.publicExponent)", "'010001'");
    shouldBeUndefined("keyPair.publicKey.algorithm.hash");
    shouldBe("keyPair.privateKey.type", "'private'");
    shouldBe("keyPair.privateKey.extractable", "false");
    shouldBe("keyPair.privateKey.algorithm.name", "'RSA-OAEP'");
    shouldBe("keyPair.privateKey.algorithm.modulusLength", "2048");
    shouldBe("bytesToHexString(keyPair.privateKey.algorithm.publicExponent)", "'010001'");
    shouldBeUndefined("keyPair.privateKey.algorithm.hash");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
